Move DynamicJniTypeManager to its own separate file
